Put basic webpack configuration loading in the WebpackPlugin.
