+++ apps/heft/src/plugins/WebpackWarningPlugin.ts
@@ -19,11 +19,8 @@ export class WebpackWarningPlugin implements IHeftPlugin {
       build.hooks.bundle.tap(PLUGIN_NAME, (bundle: IBundleSubstage) => {
         bundle.hooks.run.tapPromise(PLUGIN_NAME, async () => {
           let hasWebpackPlugin: boolean = false;
-          let hasBasicConfigureWebpackPlugin: boolean = false;
           for (const tap of bundle.hooks.run.taps) {
-            if (tap.name === 'BasicConfigureWebpackPlugin') {
-              hasBasicConfigureWebpackPlugin = true;
-            } else if (tap.name === 'WebpackPlugin') {
+            if (tap.name === 'WebpackPlugin') {
               hasWebpackPlugin = true;
             }
           }
@@ -32,8 +29,7 @@ export class WebpackWarningPlugin implements IHeftPlugin {
             heftSession,
             heftConfiguration,
             !!bundle.properties.webpackConfiguration,
-            hasWebpackPlugin,
-            hasBasicConfigureWebpackPlugin
+            hasWebpackPlugin
           );
         });
       });
@@ -44,27 +40,22 @@ export class WebpackWarningPlugin implements IHeftPlugin {
     heftSession: HeftSession,
     heftConfiguration: HeftConfiguration,
     webpackConfigIsProvided: boolean,
-    hasWebpackPlugin: boolean,
-    hasBasicConfigureWebpackPlugin: boolean
+    hasWebpackPlugin: boolean
   ): Promise<void> {
-    if (hasWebpackPlugin && hasBasicConfigureWebpackPlugin) {
-      // If we have both plugins, we don't need to check for anything else
+    if (hasWebpackPlugin) {
+      // If we have the plugin, we don't need to check anything else
       return;
     }
 
-    if (webpackConfigIsProvided && hasWebpackPlugin) {
-      // If the webpack config is already provided by some other plugin, we don't have to care about the
-      // BasicConfigureWebpackPlugin
-      return;
-    }
-
-    if (webpackConfigIsProvided && !hasWebpackPlugin) {
+    if (webpackConfigIsProvided) {
       const logger: ScopedLogger = heftSession.requestScopedLogger(PLUGIN_NAME);
       logger.emitWarning(
         new Error(
-          'A webpack configuration is provided, but the webpack plugin is missing. ' +
-            'You need to include the @rushstack/heft-webpack4-plugin plugin package ' +
-            'and reference lib/WebpackPlugin in config/heft.json.'
+          'Your project appears to have a Webpack configuration generated by a plugin, ' +
+            'but the Heft plugin for Webpack is not enabled. To fix this, you can add ' +
+            '"@rushstack/heft-webpack4-plugin" to your package.json devDependencies and use ' +
+            'config/heft.json to load it. For details, see this documentation: ' +
+            'https://rushstack.io/pages/heft_tasks/webpack/'
         )
       );
       return;
@@ -76,25 +67,15 @@ export class WebpackWarningPlugin implements IHeftPlugin {
     );
     if (webpackConfigFileExists) {
       const logger: ScopedLogger = heftSession.requestScopedLogger(PLUGIN_NAME);
-      if (hasWebpackPlugin && !hasBasicConfigureWebpackPlugin) {
-        logger.emitWarning(
-          new Error(
-            `A ${webpackConfigFilename} file exists in this project ` +
-              'but the BasicConfigureWebpackPlugin plugin is missing. ' +
-              'You probably want to include the @rushstack/heft-webpack4-plugin plugin package ' +
-              'and reference lib/BasicConfigureWebpackPlugin in config/heft.json.'
-          )
-        );
-      } else {
-        logger.emitWarning(
-          new Error(
-            `A ${webpackConfigFilename} file exists in this project ` +
-              'but the BasicConfigureWebpackPlugin plugin is missing. ' +
-              'You probably want to include the @rushstack/heft-webpack4-plugin plugin package ' +
-              'and reference lib/BasicConfigureWebpackPlugin and lib/WebpackPlugin in config/heft.json.'
-          )
-        );
-      }
+      logger.emitWarning(
+        new Error(
+          `A ${webpackConfigFilename} file exists in this project ` +
+            'but the Heft plugin for Webpack is not enabled. To fix this, you can add ' +
+            '"@rushstack/heft-webpack4-plugin" to your package.json devDependencies and use ' +
+            'config/heft.json to load it. For details, see this documentation: ' +
+            'https://rushstack.io/pages/heft_tasks/webpack/'
+        )
+      );
     }
   }
 }
