+++ apps/heft/src/stages/BuildStage.ts
@@ -131,6 +131,7 @@ export interface IBuildStageProperties {
   maxOldSpaceSize?: string;
   watchMode: boolean;
   serveMode: boolean;
+  webpackStats?: unknown;
 }
 
 /**
