+++ tutorials/heft-webpack-basic-tutorial/config/heft.json
@@ -40,18 +40,7 @@
       /**
        * The path to the plugin package.
        */
-      "plugin": "@rushstack/heft-webpack4-plugin/lib/WebpackPlugin"
-
-      /**
-       * An optional object that provides additional settings that may be defined by the plugin.
-       */
-      // "options": { }
-    },
-    {
-      /**
-       * The path to the plugin package.
-       */
-      "plugin": "@rushstack/heft-webpack4-plugin/lib/BasicConfigureWebpackPlugin"
+      "plugin": "@rushstack/heft-webpack4-plugin"
 
       /**
        * An optional object that provides additional settings that may be defined by the plugin.
